requirements.txt
streamlit==1.39.0
pandas==2.2.2
gspread==6.1.4
google-auth==2.35.0
plotly==5.24.1
openpyxl==3.1.5
requests==2.32.3
python-dotenv==1.0.1
kaleido==0.2.1
retrying==1.3.4


